requirements.txt
fastapi==0.118.0
httpx==0.28.1
pyyaml==6.0.2
pytest==8.4.1
uvicorn==0.37.0
